+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@
 ## v2.1.1 (2022-01-28)
 
 - Added option `justNow` that prints `Just now` if time is within 60 minutes;
+- Added option `noSuffix` that removes `ago` from the printed result;
 
 ----